+++ example/main.js
@@ -30,12 +30,15 @@ import {Provider, connect} from 'react-redux';
 import autobind from 'autobind-decorator';
 
 import MapboxGLMap from 'react-map-gl';
+import {FPSStats} from 'react-stats';
 import LayerSelector from './layer-selector';
 import LayerInfo from './layer-info';
 
 import * as request from 'd3-request';
 import LAYER_CATEGORIES, {DEFAULT_ACTIVE_LAYERS} from './layer-examples';
+
 import DeckGL from '../src/react/deckgl';
+import {ReflectionEffect} from '../src/experimental';
 
 // ---- Default Settings ---- //
 /* eslint-disable no-process-env */
@@ -214,8 +217,8 @@ function pointsToLines(points) {
   return points.map((point, i) => {
     if (i === points.length - 1) {
       return {
-        sourcePosition: [0, 0],
-        targetPosition: [0, 0],
+        sourcePosition: [0, 0, 0],
+        targetPosition: [0, 0, 0],
         color: [35, 81, 128]
       };
     }
@@ -224,8 +227,16 @@ function pointsToLines(points) {
     const target = points[i + 1];
 
     return {
-      sourcePosition: source.position,
-      targetPosition: target.position,
+      sourcePosition: [
+        source.position[0],
+        source.position[1],
+        Math.random() * 1000
+      ],
+      targetPosition: [
+        target.position[0],
+        target.position[1],
+        Math.random() * 1000
+      ],
       color: [0, 0, 255]
     };
   });
@@ -244,6 +255,8 @@ class ExampleApp extends React.Component {
       hoverChoropleth: null,
       clickItem: null
     };
+
+    this._effects = [new ReflectionEffect()];
   }
 
   componentWillMount() {
@@ -443,6 +456,7 @@ class ExampleApp extends React.Component {
         {...mapViewState}
         onWebGLInitialized={ this._onWebGLInitialized }
         layers={this._renderExamples()}
+        effects={this._effects}
       />
     );
   }
@@ -460,6 +474,7 @@ class ExampleApp extends React.Component {
         { ...mapViewState }
         onChangeViewport={this._handleViewportChanged}>
         { this._renderOverlay() }
+        <FPSStats isActive/>
       </MapboxGLMap>
     );
   }